requirements.txt
PyYAML
colorama
hypothesis
jinja2
numpy
pandas
xarray
pytest
pytest-cov
pytest-sugar
pytest-xdist
sphinx
breathe
sphinx_rtd_theme
mock
parsimonious
-e external/hacked_packages/pytest-cpp-0.4
